Refactor: Split DashboardCourse.tsx into smaller components
This commit refactors `DashboardCourse.tsx` into smaller, more manageable components to improve code organization and readability. Unused imports and files are also removed.
